moved code out of AST
